// src/api/dto/user.dto.ts
export class UserDTO {
    userId: number | undefined;
    name: string | undefined;
    age: number | undefined;
    email: string | undefined;
}

export class CreateUserDTO {
    name: string | undefined;
    age: number | undefined;
    email: string | undefined;
}

export class User {
    userId: number | undefined;
    name: string | undefined;
    age: number | undefined;
    email: string | undefined;
}